Tests cleanup

+++ DFCache Tests/Test Suites/TDFCache+UIImage.m
@@ -27,7 +27,7 @@
     
     static NSUInteger _index = 0;
     
-    NSString *cacheName = [NSString stringWithFormat:@"_dt_testcase_%i", _index];
+    NSString *cacheName = [NSString stringWithFormat:@"_dt_testcase_%lu", (unsigned long)_index];
     _cache = [[DFCache alloc] initWithName:cacheName];
     _index++;
 }
